LayoutTests/editing/pasteboard/paste-newline-in-all-selected-textarea.html

<!DOCTYPE html>
<html>
<body>
<textarea id="test" cols=5 rows=4>
This test verifies that pasting a newline in a textarea with all text selected
produces a newline.

</textarea>
<script src="../../resources/js-test.js"></script>
<script>
var textarea = document.getElementById('test');
textarea.focus();

// copy a new line
textarea.setSelectionRange(99, 100);
document.execCommand('Copy');

document.execCommand('SelectAll');
document.execCommand('Paste');
shouldBeEqualToString('textarea.value', '\n');
</script>
</body>
</html>
