third_party/blink/web_tests/editing/pasteboard/paste-newline-in-all-selected-textarea-expected.txt

PASS textarea.value is "\n"
PASS successfullyParsed is true

TEST COMPLETE
